disk_cache: Minimize in memory data to 2 bits
This CL upgrades the simple index in CacheStorage to use only 2 bits for
in memory data and others for entry size. With this change, now the
entry size in EntryMetadata can take up to 256GB.

Note that this change is not enough to support such large files in cache
storage yet. The work will be done in the following cls.

Bug: 391398191
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6490934
Commit-Queue: Eriko Kurimoto <elkurin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1459965}

+++ net/disk_cache/simple/simple_index_file_unittest.cc
@@ -148,9 +148,7 @@ TEST_F(SimpleIndexFileTest, Serialize) {
       static_cast<uint64_t>(kNumHashes), 456);
   for (size_t i = 0; i < kNumHashes; ++i) {
     uint64_t hash = kHashes[i];
-    // TODO(eroman): Should restructure the test so no casting here (and same
-    //               elsewhere where a hash is cast to an entry size).
-    metadata_entries[i] = EntryMetadata(Time(), static_cast<uint32_t>(hash));
+    metadata_entries[i] = EntryMetadata(Time(), hash);
     metadata_entries[i].SetInMemoryData(static_cast<uint8_t>(i));
     SimpleIndex::InsertInEntrySet(hash, metadata_entries[i], &entries);
   }
@@ -190,8 +188,7 @@ TEST_F(SimpleIndexFileTest, SerializeAppCache) {
       static_cast<uint64_t>(kNumHashes), 456);
   for (size_t i = 0; i < kNumHashes; ++i) {
     uint64_t hash = kHashes[i];
-    metadata_entries[i] =
-        EntryMetadata(kTrailerPrefetches[i], static_cast<uint32_t>(hash));
+    metadata_entries[i] = EntryMetadata(kTrailerPrefetches[i], hash);
     metadata_entries[i].SetInMemoryData(static_cast<uint8_t>(i));
     SimpleIndex::InsertInEntrySet(hash, metadata_entries[i], &entries);
   }
@@ -229,8 +226,7 @@ TEST_F(SimpleIndexFileTest, ReadV8Format) {
 
   SimpleIndex::EntrySet entries;
   for (size_t i = 0; i < kNumHashes; ++i) {
-    metadata_entries[i] =
-        EntryMetadata(base::Time::Now(), static_cast<uint32_t>(kHashes[i]));
+    metadata_entries[i] = EntryMetadata(base::Time::Now(), kHashes[i]);
     metadata_entries[i].SetInMemoryData(static_cast<uint8_t>(i));
     SimpleIndex::InsertInEntrySet(kHashes[i], metadata_entries[i], &entries);
   }
@@ -269,8 +265,7 @@ TEST_F(SimpleIndexFileTest, ReadV8FormatAppCache) {
 
   SimpleIndex::EntrySet entries;
   for (size_t i = 0; i < kNumHashes; ++i) {
-    metadata_entries[i] =
-        EntryMetadata(base::Time::Now(), static_cast<uint32_t>(kHashes[i]));
+    metadata_entries[i] = EntryMetadata(base::Time::Now(), kHashes[i]);
     metadata_entries[i].SetInMemoryData(static_cast<uint8_t>(i));
     SimpleIndex::InsertInEntrySet(kHashes[i], metadata_entries[i], &entries);
   }
